Fix tests, refactor
