+++ jabkit/src/main/java/org/jabref/cli/Convert.java
@@ -2,11 +2,19 @@ package org.jabref.cli;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.Callable;
 
+import org.jabref.logic.exporter.Exporter;
+import org.jabref.logic.exporter.ExporterFactory;
 import org.jabref.logic.importer.ParserResult;
+import org.jabref.logic.journals.JournalAbbreviationRepository;
+import org.jabref.logic.l10n.Localization;
+import org.jabref.model.database.BibDatabaseContext;
 
+import com.airhacks.afterburner.injection.Injector;
+import com.google.common.base.Throwables;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -43,10 +51,52 @@ public class Convert implements Callable<Integer> {
 
         Optional<ParserResult> pr = kitCommandLine.importFile(inputFile, inputFormat);
         if (pr.isPresent()) {
-            kitCommandLine.exportFile(pr.get(), outputFile, outputFormat);
+            exportFile(pr.get(), outputFile, outputFormat);
         } else {
             LOGGER.error("Unable to export input file {}", inputFile);
         }
         return 0;
+    }
+
+    protected void exportFile(ParserResult pr, Path outputFile, String format) {
+        if (pr == null || outputFile == null) {
+            return;
+        }
+
+        if (pr.isInvalid()) {
+            System.err.println(Localization.lang("The output option depends on a valid import option."));
+            return;
+        }
+
+        if ("bibtex".equalsIgnoreCase(format)) {
+            kitCommandLine.saveDatabase(pr.getDatabase(), outputFile);
+            return;
+        }
+
+        // This signals that the latest import should be stored in the given
+        // format to the given file.
+        Path path = pr.getPath().get().toAbsolutePath();
+        BibDatabaseContext databaseContext = pr.getDatabaseContext();
+        databaseContext.setDatabasePath(path);
+        List<Path> fileDirForDatabase = databaseContext
+                .getFileDirectories(kitCommandLine.cliPreferences.getFilePreferences());
+        System.out.println(Localization.lang("Exporting %0", outputFile));
+        ExporterFactory exporterFactory = ExporterFactory.create(kitCommandLine.cliPreferences);
+        Optional<Exporter> exporter = exporterFactory.getExporterByName(format);
+        if (exporter.isEmpty()) {
+            System.err.println(Localization.lang("Unknown export format %0", format));
+        } else {
+            // We have an exporter:
+            try {
+                exporter.get().export(
+                        pr.getDatabaseContext(),
+                        outputFile,
+                        pr.getDatabaseContext().getDatabase().getEntries(),
+                        fileDirForDatabase,
+                        Injector.instantiateModelOrService(JournalAbbreviationRepository.class));
+            } catch (Exception ex) {
+                System.err.println(Localization.lang("Could not export file '%0' (reason: %1)", outputFile, Throwables.getStackTraceAsString(ex)));
+            }
+        }
     }
 }
